clean up prototypes and split concerns into files
