+++ src/main.c
@@ -1,9 +1,6 @@
 #include <cub3D.h>
 
-void	player_movement(void *arg);
-void	put_player(t_window *window, int color);
-
-void	escape_handler(void *arg)
+static void	escape_handler(void *arg)
 {
 	mlx_t	*mlx;
 
@@ -12,7 +9,7 @@ void	escape_handler(void *arg)
 		mlx_close_window(mlx);
 }
 
-t_window	*setup_window_struct(t_map *map)
+static t_window	*setup_window_struct(t_map *map)
 {
 	t_window	*window;
 
@@ -31,31 +28,6 @@ t_window	*setup_window_struct(t_map *map)
 	window->player->dir->y = 10;
 	window->player->dir->z = HEIGHT / 2;
 	return (window);
-}
-
-void	free_map(t_cube_type **cube)
-{
-	int	i;
-
-	i = 0;
-	while (cube[i])
-	{
-		free(cube[i]);
-		i++;
-	}
-	free(cube);
-}
-
-void	free_window_struct(t_window *window)
-{
-	mlx_delete_image(window->mlx, window->img);
-	mlx_terminate(window->mlx);
-	free_map(window->map->cubes);
-	free(window->player->dir);
-	free(window->player->pos);
-	free(window->player);
-	free(window->map);
-	free(window);
 }
 
 int	main(int argc, char **argv)
